Improved Graph.Compare. Bumped to v0.8.22
Graph.Compare API has changed. Now the metrics can be compared invidividually each with their own weight. Also, all weights are defaulted to 0, so now you only need to specify the weights that you are interested in. The others will be ignored.
Dictionary.ValueatKey now behaves like the .get method for python dictionaries and returns the default value if it cannot find the key/value in the dictionary.
Bumped to version v0.8.22